+++ package.json
@@ -106,6 +106,7 @@
         "@babel/preset-env": "^7.14.4",
         "@babel/preset-typescript": "^7.13.0",
         "@types/clone": "^2.1.0",
+        "@types/express-fileupload": "^1.1.6",
         "@types/node": "^15.12.1",
         "@types/webpack": "^5.28.0",
         "@types/webpack-livereload-plugin": "^2.3.2",
