+++ schemas/tournamentData.json
@@ -1,5 +1,39 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema",
+
+    "definitions": {
+        "player": {
+            "type": "object",
+            "properties": {
+                "name": {
+                    "type": "string"
+                }
+            },
+            "additionalProperties": false
+        },
+        "team": {
+            "type": "object",
+            "properties": {
+                "id": {
+                    "type": "string"
+                },
+                "name": {
+                    "type": "string"
+                },
+                "logoUrl": {
+                    "type": "string"
+                },
+                "players": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/player"
+                    }
+                }
+            },
+            "additionalProperties": false
+        }
+    },
+
     "type": "object",
     "properties": {
         "meta": {
@@ -13,23 +47,7 @@
         "data": {
             "type": "array",
             "items": {
-                "type": "object",
-                "properties": {
-                    "id": {
-                        "type": "string"
-                    },
-                    "name": {
-                        "type": "string"
-                    },
-                    "players": {
-                        "type": "array",
-                        "items": {
-                            "name": {
-                                "type": "string"
-                            }
-                        }
-                    }
-                }
+                "$ref": "#/definitions/team"
             }
         }
     },
