+++ src/helpers/splatoonData.ts
@@ -1,0 +1,38 @@
+export const splatStages = [
+    'Ancho-V Games',
+    'Arowana Mall',
+    'Blackbelly Skatepark',
+    'Camp Triggerfish',
+    'Goby Arena',
+    'Humpback Pump Track',
+    'Inkblot Art Academy',
+    'Kelp Dome',
+    'MakoMart',
+    'Manta Maria',
+    'Moray Towers',
+    'Musselforge Fitness',
+    'New Albacore Hotel',
+    'Piranha Pit',
+    'Port Mackerel',
+    'Shellendorf Institute',
+    'Shifty Station',
+    'Snapper Canal',
+    'Starfish Mainstage',
+    'Sturgeon Shipyard',
+    'The Reef',
+    'Wahoo World',
+    'Walleye Warehouse',
+    'Skipper Pavilion',
+    'Unknown Stage'
+];
+splatStages.sort();
+
+export const splatModes = [
+    'Clam Blitz',
+    'Tower Control',
+    'Rainmaker',
+    'Splat Zones',
+    'Turf War',
+    'Unknown Mode'
+];
+splatModes.sort();
